test/KöttätandeVäxtTest.java
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class KöttätandeVäxtTest {

    //vanligtvis ska strukturen i filerna spegla sin sourcecode, men pga jag bara testar 2 klasser så har jag enbart dessa som "testfiler"

    Plantor bert = new KöttätandeVäxt(0.4,"bert");
    Plantor astrid = new KöttätandeVäxt(2, "astrid");
    Plantor meatloaf = new KöttätandeVäxt(0.7, "meatloaf");


    @Test
    void calculateVätskaMetodTest() {
       // assert astrid.getVätskebehov() == 1;   //ska ge error, vilket det gör
        bert.CalculateVätska();      //calculatevätska på berg, ska vara 0.1 + 0.2*0.4 = 0.18
        astrid.CalculateVätska();    //calculatevätska på astrid, ska vara 0.1 + 0.2*2 = 0.5
        meatloaf.CalculateVätska();  //calculatevätska på meatloaf, ska vara 0.1 + 0.2*0.7 = 0.24

        //testar att det stämmer!
        // assert bert.getVätskebehov() == 0.18;    (något fel på den här?)
        assert astrid.getVätskebehov() == 0.5;       //resten funkar
        assert meatloaf.getVätskebehov() == 0.24;


        bert.setVätskebehov(2);
        astrid.setVätskebehov(1);
        meatloaf.setVätskebehov(3);


        //kollar så det inte är något fel på setvätskebehov metoden, den metoden som används i CalculateVätska!
        assert bert.getVätskebehov() == 2;
        assert astrid.getVätskebehov() == 1;
        assert meatloaf.getVätskebehov() == 3;



        /* Jag fick error när jag testade assert bert.getVätskebehov() == 0.18, blev förvånad varför och lyckades inte lista ut det först... trodde det hade något med
        decimaler och göra fast det är ologiskt pga 0.24 stämmer ju som meatloaf hade! sen testade jag att ändra i huvudprogrammet till att meatloaf är 0.4 meter i höjd
        och fick då ut att svaret är 0.1800000000000002 ????
        lade till avrundningsmetod för string i utskriften för att lösa detta
        String.format("%.2f", this.getVätskebehov())
        verkligen konstigt att 0.1+0.2*0.4 får problem att räkna ut det till 0.18 men att 0.1+0.2*0.7 fungerar fint?
        */
    }
}